Restructures navigation icon
